Move the QueryBoundary and Tab groups to their own entrypoint
They're big enough to warrant their own entrypoints by now.
