+++ src/root/components/index.ts
@@ -4,8 +4,6 @@ export { default as ExternalLink } from "src/root/components/ExternalLink";
 export { default as FileInput, FileType } from "src/root/components/FileInput";
 export { default as FileInputList } from "src/root/components/FileInputList";
 export { default as Page } from "src/root/components/Page";
-export { default as QueryBoundaryItemWrapper } from "src/root/components/QueryBoundaryItemWrapper";
-export { default as QueryBoundaryListWrapper } from "src/root/components/QueryBoundaryListWrapper";
 export { default as SkeletonRow } from "src/root/components/SkeletonRow";
 export { default as SwitchWithIcons } from "src/root/components/SwitchWithIcons";
 export { default as ThemeToggle } from "src/root/components/ThemeToggle";
@@ -14,7 +12,5 @@ export type { CollapsibleItemProps } from "src/root/components/CollapsibleItem";
 export type { ExternalLinkProps } from "src/root/components/ExternalLink";
 export type { FileInputProps } from "src/root/components/FileInput";
 export type { FileInputListProps } from "src/root/components/FileInputList";
-export type { QueryBoundaryItemWrapperProps } from "src/root/components/QueryBoundaryItemWrapper";
-export type { QueryBoundaryListWrapperProps } from "src/root/components/QueryBoundaryListWrapper";
 export type { SkeletonRowProps } from "src/root/components/SkeletonRow";
 export type { SwitchWithIconsProps } from "src/root/components/SwitchWithIcons";
