--- src/root/groups/QueryBoundary/creators/index.ts
@@ -1,19 +1,0 @@
-export { default as createBaseQueryBoundary } from "src/root/groups/QueryBoundary/creators/createBaseQueryBoundary";
-export { default as createItemQueryBoundary } from "src/root/groups/QueryBoundary/creators/createItemQueryBoundary";
-export { default as createListQueryBoundary } from "src/root/groups/QueryBoundary/creators/createListQueryBoundary";
-
-export type {
-  QueryBase,
-  DefaultQueryBoundaryComponentsBase,
-  CreateBaseQueryBoundaryParameters,
-} from "src/root/groups/QueryBoundary/creators/createBaseQueryBoundary";
-export type {
-  QueryItem,
-  DefaultQueryBoundaryItemComponents,
-  CreateItemQueryBoundaryParameters,
-} from "src/root/groups/QueryBoundary/creators/createItemQueryBoundary";
-export type {
-  QueryList,
-  DefaultQueryBoundaryListComponents,
-  CreateListQueryBoundaryParameters,
-} from "src/root/groups/QueryBoundary/creators/createListQueryBoundary";
