--- src/root/groups/Tab/index.ts
@@ -1,10 +1,0 @@
-export { default as createTabGroup } from "src/root/groups/Tab/createTabGroup";
-export { default as TabItem } from "src/root/groups/Tab/TabItem";
-export { default as TabList } from "src/root/groups/Tab/TabList";
-export { default as TabPanel } from "src/root/groups/Tab/TabPanel";
-export { default as TabProvider, useTabContext } from "src/root/groups/Tab/TabProvider";
-
-export type { TabItemProps } from "src/root/groups/Tab/TabItem";
-export type { TabListProps } from "src/root/groups/Tab/TabList";
-export type { TabPanelProps } from "src/root/groups/Tab/TabPanel";
-export type { TabProviderProps, TabContextValue } from "src/root/groups/Tab/TabProvider";
