--- src/root/groups/index.ts
@@ -1,2 +1,0 @@
-export * from "src/root/groups/QueryBoundary";
-export * from "src/root/groups/Tab";
