+++ tsdown.config.ts
@@ -1,31 +1,36 @@
 import type { UserConfig } from "tsdown";
 
+const baseConfig: UserConfig = {
+  format: ["esm"],
+  dts: true,
+  deps: {
+    neverBundle: ["react"],
+  },
+  clean: true,
+  minify: true,
+  sourcemap: true,
+  fixedExtension: false,
+};
+
 const config: Array<UserConfig> = [
   {
     entry: ["src/root/index.ts"],
-    format: ["esm", "cjs"],
-    dts: true,
-    deps: {
-      neverBundle: ["react"],
-    },
-    clean: true,
-    minify: true,
-    sourcemap: true,
-    fixedExtension: false,
+    outDir: "dist",
   },
   {
-    entry: ["src/v7/index.ts"],
-    format: ["esm", "cjs"],
-    outDir: "dist/v7",
-    dts: true,
-    deps: {
-      neverBundle: ["react"],
-    },
-    clean: true,
-    minify: true,
-    sourcemap: true,
-    fixedExtension: false,
+    entry: ["src/QueryBoundary/index.ts"],
+    outDir: "dist/QueryBoundary",
   },
-];
+  {
+    entry: ["src/Tabs/index.ts"],
+    outDir: "dist/Tabs",
+  },
+].map(({ entry, outDir }) => {
+  return {
+    entry,
+    outDir,
+    ...baseConfig,
+  };
+});
 
 export default config;
